requirements.txt
Django==4.1.4
djangorestframework==3.14.0
django-cors-headers==4.0.0
django-celery-results==2.5.1
django-celery-beat==2.5.0
python-crontab==2.7.1
psycopg2==2.9.6
celery
finnhub-python
redis
